feat: extract storage preset components into routes
