+++ packages/devtools/client/src/entries/client/routes/storage/preset/[id]/page.module.scss
@@ -1,0 +1,5 @@
+.text-ellipsis {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
