+++ packages/devtools/client/src/entries/client/routes/storage/preset/page.module.scss
@@ -53,9 +53,3 @@
 .divider {
   border-top: solid 1px var(--gray-4);
 }
-
-.text-ellipsis {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
