+++ packages/devtools/client/src/entries/client/routes/storage/preset/page.tsx
@@ -1,0 +1,7 @@
+import { FC } from 'react';
+
+const Page: FC = () => {
+  return <div>hello</div>;
+};
+
+export default Page;
