+++ tests/integration/devtools/modern.runtime.json
@@ -57,15 +57,8 @@
       "id": "YG5V4dtjmRqdm705f6Cbh"
     },
     {
-      "name": "New Preset yVVoN0",
-      "id": "yVVoN0WcCBhaXm6x7Uwix",
-      "cookie": {},
-      "localStorage": {},
-      "sessionStorage": {}
-    },
-    {
-      "name": "New Preset 6m9DXm",
-      "id": "6m9DXmcT88xh0R7WUpO-u",
+      "name": "New Preset XBBHSE",
+      "id": "XBBHSEx2DfWBj3mJkq-OY",
       "cookie": {},
       "localStorage": {},
       "sessionStorage": {}
